+++ src/algorithm/boundary_condition/standard_bc.hpp
@@ -3,7 +3,7 @@
 #include "../BootesArray.hpp"
 
 
-void standard_boundary_condition_x1i(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void standard_boundary_condition_x1i(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                             int &x2s, int &x2l, int &ng2,
                                                             int &x3s, int &x3l, int &ng3){
     // standard inflow/outflow boundary, simply copy the values from active zone to ghost zone.
@@ -20,7 +20,7 @@ void standard_boundary_condition_x1i(BootesArray<float> &quan, int &x1s, int &x1
 }
 
 
-void standard_boundary_condition_x1o(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void standard_boundary_condition_x1o(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                             int &x2s, int &x2l, int &ng2,
                                                             int &x3s, int &x3l, int &ng3){
     // standard inflow/outflow boundary, simply copy the values from active zone to ghost zone.
@@ -37,7 +37,7 @@ void standard_boundary_condition_x1o(BootesArray<float> &quan, int &x1s, int &x1
 }
 
 
-void standard_boundary_condition_x2i(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void standard_boundary_condition_x2i(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                             int &x2s, int &x2l, int &ng2,
                                                             int &x3s, int &x3l, int &ng3){
     // standard inflow/outflow boundary, simply copy the values from active zone to ghost zone.
@@ -53,7 +53,7 @@ void standard_boundary_condition_x2i(BootesArray<float> &quan, int &x1s, int &x1
     }
 }
 
-void standard_boundary_condition_x2o(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void standard_boundary_condition_x2o(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                             int &x2s, int &x2l, int &ng2,
                                                             int &x3s, int &x3l, int &ng3){
     // standard inflow/outflow boundary, simply copy the values from active zone to ghost zone.
@@ -70,7 +70,7 @@ void standard_boundary_condition_x2o(BootesArray<float> &quan, int &x1s, int &x1
 }
 
 
-void standard_boundary_condition_x3i(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void standard_boundary_condition_x3i(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                             int &x2s, int &x2l, int &ng2,
                                                             int &x3s, int &x3l, int &ng3){
     // standard inflow/outflow boundary, simply copy the values from active zone to ghost zone.
@@ -87,7 +87,7 @@ void standard_boundary_condition_x3i(BootesArray<float> &quan, int &x1s, int &x1
 }
 
 
-void standard_boundary_condition_x3o(BootesArray<float> &quan, int &x1s, int &x1l, int &ng1,
+void standard_boundary_condition_x3o(BootesArray<double> &quan, int &x1s, int &x1l, int &ng1,
                                                             int &x2s, int &x2l, int &ng2,
                                                             int &x3s, int &x3l, int &ng3){
     // standard inflow/outflow boundary, simply copy the values from active zone to ghost zone.
